HelVM.HelPA.Assemblers.Frontend.EIR.DSL

Documentation

mov :: (EIR a, MonadDSL a m) => ImmediateORRegister -> Register -> m ()

add :: (EIR a, MonadDSL a m) => ImmediateORRegister -> Register -> m ()

sub :: (EIR a, MonadDSL a m) => ImmediateORRegister -> Register -> m ()

load :: (EIR a, MonadDSL a m) => ImmediateORRegister -> Register -> m ()

store :: (EIR a, MonadDSL a m) => Register -> ImmediateORRegister -> m ()

putc :: (EIR a, MonadDSL a m) => ImmediateORRegister -> m ()

jeq :: (EIR a, MonadDSL a m) => ImmediateORRegister -> Register -> Identifier -> m ()

jne :: (EIR a, MonadDSL a m) => ImmediateORRegister -> Register -> Identifier -> m ()

jlt :: (EIR a, MonadDSL a m) => ImmediateORRegister -> Register -> Identifier -> m ()

jgt :: (EIR a, MonadDSL a m) => ImmediateORRegister -> Register -> Identifier -> m ()

jle :: (EIR a, MonadDSL a m) => ImmediateORRegister -> Register -> Identifier -> m ()

jge :: (EIR a, MonadDSL a m) => ImmediateORRegister -> Register -> Identifier -> m ()

eq :: (EIR a, MonadDSL a m) => ImmediateORRegister -> Register -> m ()

ne :: (EIR a, MonadDSL a m) => ImmediateORRegister -> Register -> m ()

lt :: (EIR a, MonadDSL a m) => ImmediateORRegister -> Register -> m ()

gt :: (EIR a, MonadDSL a m) => ImmediateORRegister -> Register -> m ()

le :: (EIR a, MonadDSL a m) => ImmediateORRegister -> Register -> m ()

ge :: (EIR a, MonadDSL a m) => ImmediateORRegister -> Register -> m ()

class EIR a where

Minimal complete definition

movi, movr, addr, subr, loadr, storer, putcr, getc, jeqr, jner, jltr, jgtr, jler, jger, jmp, eqr, ner, ltr, gtr, ler, ger, mark

Methods

movi :: MonadDSL a m => Immediate -> Register -> m ()

movr :: MonadDSL a m => Register -> Register -> m ()

addi :: MonadDSL a m => Immediate -> Register -> m ()

addr :: MonadDSL a m => Register -> Register -> m ()

subi :: MonadDSL a m => Immediate -> Register -> m ()

subr :: MonadDSL a m => Register -> Register -> m ()

loadi :: MonadDSL a m => Immediate -> Register -> m ()

loadr :: MonadDSL a m => Register -> Register -> m ()

storei :: MonadDSL a m => Register -> Immediate -> m ()

storer :: MonadDSL a m => Register -> Register -> m ()

putci :: MonadDSL a m => Immediate -> m ()

putcr :: MonadDSL a m => Register -> m ()

getc :: MonadDSL a m => Register -> m ()

jeqi :: MonadDSL a m => Immediate -> Register -> Identifier -> m ()

jeqr :: MonadDSL a m => Register -> Register -> Identifier -> m ()

jnei :: MonadDSL a m => Immediate -> Register -> Identifier -> m ()

jner :: MonadDSL a m => Register -> Register -> Identifier -> m ()

jlti :: MonadDSL a m => Immediate -> Register -> Identifier -> m ()

jltr :: MonadDSL a m => Register -> Register -> Identifier -> m ()

jgti :: MonadDSL a m => Immediate -> Register -> Identifier -> m ()

jgtr :: MonadDSL a m => Register -> Register -> Identifier -> m ()

jlei :: MonadDSL a m => Immediate -> Register -> Identifier -> m ()

jler :: MonadDSL a m => Register -> Register -> Identifier -> m ()

jgei :: MonadDSL a m => Immediate -> Register -> Identifier -> m ()

jger :: MonadDSL a m => Register -> Register -> Identifier -> m ()

jmp :: MonadDSL a m => Identifier -> m ()

eqi :: MonadDSL a m => Immediate -> Register -> m ()

eqr :: MonadDSL a m => Register -> Register -> m ()

nei :: MonadDSL a m => Immediate -> Register -> m ()

ner :: MonadDSL a m => Register -> Register -> m ()

lti :: MonadDSL a m => Immediate -> Register -> m ()

ltr :: MonadDSL a m => Register -> Register -> m ()

gti :: MonadDSL a m => Immediate -> Register -> m ()

gtr :: MonadDSL a m => Register -> Register -> m ()

lei :: MonadDSL a m => Immediate -> Register -> m ()

ler :: MonadDSL a m => Register -> Register -> m ()

gei :: MonadDSL a m => Immediate -> Register -> m ()

ger :: MonadDSL a m => Register -> Register -> m ()

dump :: MonadDSL a m => m ()

mark :: MonadDSL a m => Identifier -> m ()

pText :: MonadDSL a m => m ()

pData :: MonadDSL a m => Maybe Natural -> m ()

pLong :: MonadDSL a m => Integer -> m ()

pString :: MonadDSL a m => Text -> m ()

pFile :: MonadDSL a m => Natural -> Identifier -> m ()

pLoc :: MonadDSL a m => Natural -> Natural -> Natural -> m ()