HelVM.HelPA.Assemblers.Frontend.ELVM.DSL

Documentation

newtype Register

Constructors

R Text

register :: Register

newtype Immediate

Constructors

I Integer

unRegister :: Register -> Text

unImmediate :: Immediate -> Integer

type ImmediateORRegister = Either Immediate Register

mov :: MonadDSL a => ImmediateORRegister -> Register -> DSL a

add :: MonadDSL a => ImmediateORRegister -> Register -> DSL a

sub :: MonadDSL a => ImmediateORRegister -> Register -> DSL a

load :: MonadDSL a => ImmediateORRegister -> Register -> DSL a

store :: MonadDSL a => Register -> ImmediateORRegister -> DSL a

putc :: MonadDSL a => ImmediateORRegister -> DSL a

jmp :: MonadDSL a => ImmediateORRegister -> Register -> DSL a

class MonadDSL a where

Minimal complete definition

movi, movr, addr, subr, loadi, loadr, storei, storer, putci, putcr, getc, jeq, jne, jlt, jgt, jle, jge, jmpi, jmpr, eqi, eqr, nei, ner, lti, ltr, gti, gtr, lei, ler, gei, ger, dump

Methods

movi :: Immediate -> Register -> DSL a

movr :: Register -> Register -> DSL a

addi :: Immediate -> Register -> DSL a

addr :: Register -> Register -> DSL a

subi :: Immediate -> Register -> DSL a

subr :: Register -> Register -> DSL a

loadi :: Immediate -> Register -> DSL a

loadr :: Register -> Register -> DSL a

storei :: Register -> Immediate -> DSL a

storer :: Register -> Register -> DSL a

putci :: Immediate -> DSL a

putcr :: Register -> DSL a

getc :: Register -> DSL a

jeq :: ImmediateORRegister -> ImmediateORRegister -> Register -> DSL a

jne :: ImmediateORRegister -> ImmediateORRegister -> Register -> DSL a

jlt :: ImmediateORRegister -> ImmediateORRegister -> Register -> DSL a

jgt :: ImmediateORRegister -> ImmediateORRegister -> Register -> DSL a

jle :: ImmediateORRegister -> ImmediateORRegister -> Register -> DSL a

jge :: ImmediateORRegister -> ImmediateORRegister -> Register -> DSL a

jmpi :: Immediate -> Register -> DSL a

jmpr :: Register -> Register -> DSL a

eqi :: Immediate -> Register -> DSL a

eqr :: Register -> Register -> DSL a

nei :: Immediate -> Register -> DSL a

ner :: Register -> Register -> DSL a

lti :: Immediate -> Register -> DSL a

ltr :: Register -> Register -> DSL a

gti :: Immediate -> Register -> DSL a

gtr :: Register -> Register -> DSL a

lei :: Immediate -> Register -> DSL a

ler :: Register -> Register -> DSL a

gei :: Immediate -> Register -> DSL a

ger :: Register -> Register -> DSL a

dump :: DSL a