HelVM.HelPA.Assemblers.Frontend.ELVM.DSL

Documentation

mov :: (ELVM a, MonadDSL a m) => ImmediateORRegister -> Register -> m ()

add :: (ELVM a, MonadDSL a m) => ImmediateORRegister -> Register -> m ()

sub :: (ELVM a, MonadDSL a m) => ImmediateORRegister -> Register -> m ()

load :: (ELVM a, MonadDSL a m) => ImmediateORRegister -> Register -> m ()

store :: (ELVM a, MonadDSL a m) => Register -> ImmediateORRegister -> m ()

putc :: (ELVM a, MonadDSL a m) => ImmediateORRegister -> m ()

jeq :: (ELVM a, MonadDSL a m) => ImmediateORRegister -> ImmediateORRegister -> Register -> m ()

jne :: (ELVM a, MonadDSL a m) => ImmediateORRegister -> ImmediateORRegister -> Register -> m ()

jlt :: (ELVM a, MonadDSL a m) => ImmediateORRegister -> ImmediateORRegister -> Register -> m ()

jgt :: (ELVM a, MonadDSL a m) => ImmediateORRegister -> ImmediateORRegister -> Register -> m ()

jle :: (ELVM a, MonadDSL a m) => ImmediateORRegister -> ImmediateORRegister -> Register -> m ()

jge :: (ELVM a, MonadDSL a m) => ImmediateORRegister -> ImmediateORRegister -> Register -> m ()

jmp :: (ELVM a, MonadDSL a m) => ImmediateORRegister -> m ()

eq :: (ELVM a, MonadDSL a m) => ImmediateORRegister -> Register -> m ()

ne :: (ELVM a, MonadDSL a m) => ImmediateORRegister -> Register -> m ()

lt :: (ELVM a, MonadDSL a m) => ImmediateORRegister -> Register -> m ()

gt :: (ELVM a, MonadDSL a m) => ImmediateORRegister -> Register -> m ()

le :: (ELVM a, MonadDSL a m) => ImmediateORRegister -> Register -> m ()

ge :: (ELVM a, MonadDSL a m) => ImmediateORRegister -> Register -> m ()

class ELVM a where

Minimal complete definition

movi, movr, addr, subr, loadr, storer, putcr, getc, jeqrr, jnerr, jltrr, jgtrr, jlerr, jgerr, jmpr, eqr, ner, ltr, gtr, ler, ger

Methods

movi :: MonadDSL a m => Immediate -> Register -> m ()

movr :: MonadDSL a m => Register -> Register -> m ()

addi :: MonadDSL a m => Immediate -> Register -> m ()

addr :: MonadDSL a m => Register -> Register -> m ()

subi :: MonadDSL a m => Immediate -> Register -> m ()

subr :: MonadDSL a m => Register -> Register -> m ()

loadi :: MonadDSL a m => Immediate -> Register -> m ()

loadr :: MonadDSL a m => Register -> Register -> m ()

storei :: MonadDSL a m => Register -> Immediate -> m ()

storer :: MonadDSL a m => Register -> Register -> m ()

putci :: MonadDSL a m => Immediate -> m ()

putcr :: MonadDSL a m => Register -> m ()

getc :: MonadDSL a m => Register -> m ()

jeqii :: MonadDSL a m => Immediate -> Immediate -> Register -> m ()

jeqir :: MonadDSL a m => Immediate -> Register -> Register -> m ()

jeqri :: MonadDSL a m => Register -> Immediate -> Register -> m ()

jeqrr :: MonadDSL a m => Register -> Register -> Register -> m ()

jneii :: MonadDSL a m => Immediate -> Immediate -> Register -> m ()

jneir :: MonadDSL a m => Immediate -> Register -> Register -> m ()

jneri :: MonadDSL a m => Register -> Immediate -> Register -> m ()

jnerr :: MonadDSL a m => Register -> Register -> Register -> m ()

jltii :: MonadDSL a m => Immediate -> Immediate -> Register -> m ()

jltir :: MonadDSL a m => Immediate -> Register -> Register -> m ()

jltri :: MonadDSL a m => Register -> Immediate -> Register -> m ()

jltrr :: MonadDSL a m => Register -> Register -> Register -> m ()

jgtii :: MonadDSL a m => Immediate -> Immediate -> Register -> m ()

jgtir :: MonadDSL a m => Immediate -> Register -> Register -> m ()

jgtri :: MonadDSL a m => Register -> Immediate -> Register -> m ()

jgtrr :: MonadDSL a m => Register -> Register -> Register -> m ()

jleii :: MonadDSL a m => Immediate -> Immediate -> Register -> m ()

jleir :: MonadDSL a m => Immediate -> Register -> Register -> m ()

jleri :: MonadDSL a m => Register -> Immediate -> Register -> m ()

jlerr :: MonadDSL a m => Register -> Register -> Register -> m ()

jgeii :: MonadDSL a m => Immediate -> Immediate -> Register -> m ()

jgeir :: MonadDSL a m => Immediate -> Register -> Register -> m ()

jgeri :: MonadDSL a m => Register -> Immediate -> Register -> m ()

jgerr :: MonadDSL a m => Register -> Register -> Register -> m ()

jmpi :: MonadDSL a m => Immediate -> m ()

jmpr :: MonadDSL a m => Register -> m ()

eqi :: MonadDSL a m => Immediate -> Register -> m ()

eqr :: MonadDSL a m => Register -> Register -> m ()

nei :: MonadDSL a m => Immediate -> Register -> m ()

ner :: MonadDSL a m => Register -> Register -> m ()

lti :: MonadDSL a m => Immediate -> Register -> m ()

ltr :: MonadDSL a m => Register -> Register -> m ()

gti :: MonadDSL a m => Immediate -> Register -> m ()

gtr :: MonadDSL a m => Register -> Register -> m ()

lei :: MonadDSL a m => Immediate -> Register -> m ()

ler :: MonadDSL a m => Register -> Register -> m ()

gei :: MonadDSL a m => Immediate -> Register -> m ()

ger :: MonadDSL a m => Register -> Register -> m ()